+++ test/once_common_structs/models/table2/generated.rs
@@ -9,7 +9,7 @@ use crate::models::common::*;
 type ConnectionType = diesel::r2d2::PooledConnection<diesel::r2d2::ConnectionManager<diesel::PgConnection>>;
 
 /// Struct representing a row in table `table2`
-#[derive(Debug, Clone, Serialize, Deserialize, Queryable, Selectable)]
+#[derive(Debug, Clone, Serialize, Deserialize, Queryable, Selectable, QueryableByName)]
 #[diesel(table_name=table2, primary_key(id))]
 pub struct Table2 {
     /// Field representing column `id`