+++ test/simple_table_no_crud/models/todos/generated.rs
@@ -5,7 +5,7 @@ use crate::schema::*;
 use serde::{Deserialize, Serialize};
 
 /// Struct representing a row in table `todos`
-#[derive(Debug, Clone, Serialize, Deserialize, Queryable, Selectable)]
+#[derive(Debug, Clone, Serialize, Deserialize, Queryable, Selectable, QueryableByName)]
 #[diesel(table_name=todos, primary_key(id))]
 pub struct Todos {
     /// Field representing column `id`